For issue #66.
Moves FlagGuard out from World.cpp and up to general usability. Also removes some stuff no longer used.
